requirements.txt
python-dateutil
jsonschema
matplotlib
pyyaml
setuptools
